Adding new namespace folder.
