+++ dataactvalidator/interfaces/errorInterface.py
@@ -1,6 +1,6 @@
 from dataactcore.models.errorModels import FileStatus, ErrorData, ErrorType, Status
 from dataactcore.models import errorInterface
-from validation_handlers.validationError import ValidationError
+from dataactvalidator.validation_handlers.validationError import ValidationError
 
 class ErrorInterface(errorInterface.ErrorInterface):
     """ Manages communication with the error database """
