+++ dataactvalidator/scripts/loadApprop.py
@@ -1,0 +1,4 @@
+from dataactvalidator.filestreaming.schemaLoader import SchemaLoader
+
+SchemaLoader.loadFields("appropriations","../tests/appropriationsFields.csv")
+SchemaLoader.loadRules("appropriations","../tests/appropriationsRules.csv")
