+++ dataactvalidator/scripts/setupTASIndexs.py
