--- scripts/tasSetup.py
@@ -1,6 +1,0 @@
-from filestreaming.tasLoader import TASLoader
-from scripts.setupTASIndexs import setupTASIndexs
-
-def loadTAS(filename):
-    TASLoader.loadFields(filename)
-    setupTASIndexs()
